chore(link-helper): move link-helper from input-field to util
